refactor(grid): Deprecate unused datagrid directive.
Resolves #554.
